+++ src/hdmf/common/resources.py
@@ -628,7 +628,7 @@ class HERD(Container):
             if entity_uri is not None:
                 entity_uri = entity.entity_uri
                 msg = 'This entity already exists. Ignoring new entity uri'
-                warn(msg)
+                warn(msg, stacklevel=2)
 
         #################
         # Validate Object
@@ -897,7 +897,7 @@ class HERD(Container):
 
     @docval({'name': 'use_categories', 'type': bool, 'default': False,
              'doc': 'Use a multi-index on the columns to indicate which category each column belongs to.'},
-            rtype=pd.DataFrame, returns='A DataFrame with all data merged into a flat, denormalized table.')
+            rtype='pandas.DataFrame', returns='A DataFrame with all data merged into a flat, denormalized table.')
     def to_dataframe(self, **kwargs):
         """
         Convert the data from the keys, resources, entities, objects, and object_keys tables
@@ -991,12 +991,21 @@ class HERD(Container):
 
     @classmethod
     @docval({'name': 'path', 'type': str, 'doc': 'The path to the zip file.'})
+    def get_zip_directory(cls, path):
+        """
+        Return the directory of the file given.
+        """
+        directory = os.path.dirname(os.path.realpath(path))
+        return directory
+
+    @classmethod
+    @docval({'name': 'path', 'type': str, 'doc': 'The path to the zip file.'})
     def from_zip(cls, **kwargs):
         """
         Method to read in zipped tsv files to populate HERD.
         """
         zip_file = kwargs['path']
-        directory = os.path.dirname(zip_file)
+        directory = cls.get_zip_directory(zip_file)
 
         with zipfile.ZipFile(zip_file, 'r') as zip:
             zip.extractall(directory)