+++ tests/unit/validator_tests/test_validate.py
@@ -116,7 +116,18 @@ class TestDateTimeInSpec(ValidatorTestBase):
                 ),
                 DatasetSpec('an example time dataset', 'isodatetime', name='datetime'),
                 DatasetSpec('an example time dataset', 'isodatetime', name='date', quantity='?'),
-                DatasetSpec('an array of times', 'isodatetime', name='time_array', dims=('num_times',), shape=(None,))
+                DatasetSpec('an array of times', 'isodatetime', name='time_array', dims=('num_times',), shape=(None,)),
+                DatasetSpec(
+                    doc='an array with compound dtype that includes an isodatetime',
+                    dtype=[
+                        DtypeSpec('x', doc='x', dtype='int'),
+                        DtypeSpec('y', doc='y', dtype='isodatetime'),
+                    ],
+                    name='cpd_array',
+                    dims=('num_times',),
+                    shape=(None,),
+                    quantity="?",
+                ),
             ],
             attributes=[AttributeSpec('attr1', 'an example string attribute', 'text')])
         return ret,
@@ -129,7 +140,15 @@ class TestDateTimeInSpec(ValidatorTestBase):
                 DatasetBuilder('data', 100, attributes={'attr2': 10}),
                 DatasetBuilder('datetime', datetime(2017, 5, 1, 12, 0, 0)),
                 DatasetBuilder('date', date(2017, 5, 1)),
-                DatasetBuilder('time_array', [datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())])
+                DatasetBuilder('time_array', [datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())]),
+                DatasetBuilder(
+                    name='cpd_array',
+                    data=[(1, datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal()))],
+                    dtype=[
+                        DtypeSpec('x', doc='x', dtype='int'),
+                        DtypeSpec('y', doc='y', dtype='isodatetime'),
+                    ],
+                ),
             ]
         )
         validator = self.vmap.get_validator('Bar')
@@ -143,7 +162,7 @@ class TestDateTimeInSpec(ValidatorTestBase):
             datasets=[
                 DatasetBuilder('data', 100, attributes={'attr2': 10}),
                 DatasetBuilder('datetime', 100),
-                DatasetBuilder('time_array', [datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())])
+                DatasetBuilder('time_array', [datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())]),
             ]
         )
         validator = self.vmap.get_validator('Bar')
@@ -152,17 +171,43 @@ class TestDateTimeInSpec(ValidatorTestBase):
         self.assertValidationError(result[0], DtypeError, name='Bar/datetime')
 
     def test_invalid_isodatetime_array(self):
-        builder = GroupBuilder('my_bar',
-                               attributes={'data_type': 'Bar', 'attr1': 'a string attribute'},
-                               datasets=[DatasetBuilder('data', 100, attributes={'attr2': 10}),
-                                         DatasetBuilder('datetime',
-                                                        datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())),
-                                         DatasetBuilder('time_array',
-                                                        datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal()))])
+        builder = GroupBuilder(
+            'my_bar',
+            attributes={'data_type': 'Bar', 'attr1': 'a string attribute'},
+            datasets=[
+                DatasetBuilder('data', 100, attributes={'attr2': 10}),
+                DatasetBuilder('datetime', datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())),
+                DatasetBuilder('time_array', datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())),
+            ],
+        )
         validator = self.vmap.get_validator('Bar')
         result = validator.validate(builder)
         self.assertEqual(len(result), 1)
         self.assertValidationError(result[0], ExpectedArrayError, name='Bar/time_array')
+
+    def test_invalid_cpd_isodatetime_array(self):
+        builder = GroupBuilder(
+            'my_bar',
+            attributes={'data_type': 'Bar', 'attr1': 'a string attribute'},
+            datasets=[
+                DatasetBuilder('data', 100, attributes={'attr2': 10}),
+                DatasetBuilder('datetime', datetime(2017, 5, 1, 12, 0, 0)),
+                DatasetBuilder('date', date(2017, 5, 1)),
+                DatasetBuilder('time_array', [datetime(2017, 5, 1, 12, 0, 0, tzinfo=tzlocal())]),
+                DatasetBuilder(
+                    name='cpd_array',
+                    data=[(1, "wrong")],
+                    dtype=[
+                        DtypeSpec('x', doc='x', dtype='int'),
+                        DtypeSpec('y', doc='y', dtype='isodatetime'),
+                    ],
+                ),
+            ],
+        )
+        validator = self.vmap.get_validator('Bar')
+        result = validator.validate(builder)
+        self.assertEqual(len(result), 1)
+        self.assertValidationError(result[0], DtypeError, name='Bar/cpd_array')
 
 
 class TestNestedTypes(ValidatorTestBase):
@@ -506,6 +551,58 @@ class Test1DArrayValidation(TestCase):
         self.assertEqual(len(results), 0)
 
     # TODO test shape validation more completely
+
+
+class TestStringDatetime(TestCase):
+
+    def test_str_coincidental_isodatetime(self):
+        """Test validation of a text spec allows a string that coincidentally matches the isodatetime format."""
+        spec_catalog = SpecCatalog()
+        spec = GroupSpec(
+            doc='A test group specification with a data type',
+            data_type_def='Bar',
+            datasets=[
+                DatasetSpec(doc='an example scalar dataset', dtype="text", name='data1'),
+                DatasetSpec(doc='an example 1D dataset', dtype="text", name='data2', shape=(None, )),
+                DatasetSpec(
+                    doc='an example 1D compound dtype dataset',
+                    dtype=[
+                        DtypeSpec('x', doc='x', dtype='int'),
+                        DtypeSpec('y', doc='y', dtype='text'),
+                    ],
+                    name='data3',
+                    shape=(None, ),
+                ),
+            ],
+            attributes=[
+                AttributeSpec(name='attr1', doc='an example scalar attribute', dtype="text"),
+                AttributeSpec(name='attr2', doc='an example 1D attribute', dtype="text", shape=(None, )),
+            ]
+        )
+        spec_catalog.register_spec(spec, 'test.yaml')
+        namespace = SpecNamespace(
+            'a test namespace', CORE_NAMESPACE, [{'source': 'test.yaml'}], version='0.1.0', catalog=spec_catalog
+        )
+        vmap = ValidatorMap(namespace)
+
+        bar_builder = GroupBuilder(
+            name='my_bar',
+            attributes={'data_type': 'Bar', 'attr1': "2023-01-01", 'attr2': ["2023-01-01"]},
+            datasets=[
+                DatasetBuilder(name='data1', data="2023-01-01"),
+                DatasetBuilder(name='data2', data=["2023-01-01"]),
+                DatasetBuilder(
+                    name='data3',
+                    data=[(1, "2023-01-01")],
+                    dtype=[
+                        DtypeSpec('x', doc='x', dtype='int'),
+                        DtypeSpec('y', doc='y', dtype='text'),
+                    ],
+                ),
+            ],
+        )
+        results = vmap.validate(bar_builder)
+        self.assertEqual(len(results), 0)
 
 
 class TestLinkable(TestCase):
@@ -1021,3 +1118,102 @@ class TestReferenceDatasetsRoundTrip(ValidatorTestBase):
             attributes={'data_type': 'Foo'}
         )
         self.runBuilderRoundTrip(foo)
+
+
+class TestEmptyDataRoundTrip(ValidatorTestBase):
+    """
+    Test the special case of empty string datasets and attributes during validation
+    """
+    def setUp(self):
+        self.filename = 'test_ref_dataset.h5'
+        super().setUp()
+
+    def tearDown(self):
+        remove_test_file(self.filename)
+        super().tearDown()
+
+    def getSpecs(self):
+        ret = GroupSpec('A test group specification with a data type',
+                        data_type_def='Bar',
+                        datasets=[DatasetSpec(name='data',
+                                              doc='an example dataset',
+                                              dtype='text',
+                                              attributes=[AttributeSpec(
+                                                  name='attr2',
+                                                  doc='an example integer attribute',
+                                                  dtype='int',
+                                                  shape=(None,))]),
+                                  DatasetSpec(name='dataInt',
+                                              doc='an example int dataset',
+                                              dtype='int',
+                                              attributes=[])
+                                  ],
+                        attributes=[AttributeSpec(name='attr1',
+                                                  doc='an example string attribute',
+                                                  dtype='text',
+                                                  shape=(None,))])
+        return (ret,)
+
+    def runBuilderRoundTrip(self, builder):
+        """Executes a round-trip test for a builder
+
+        1. First writes the builder to file,
+        2. next reads a new builder from disk
+        3. and finally runs the builder through the validator.
+        The test is successful if there are no validation errors."""
+        ns_catalog = NamespaceCatalog()
+        ns_catalog.add_namespace(self.namespace.name, self.namespace)
+        typemap = TypeMap(ns_catalog)
+        self.manager = BuildManager(typemap)
+
+        with HDF5IO(self.filename, manager=self.manager, mode='w') as write_io:
+            write_io.write_builder(builder)
+
+        with HDF5IO(self.filename, manager=self.manager, mode='r') as read_io:
+            read_builder = read_io.read_builder()
+            errors = self.vmap.validate(read_builder)
+            self.assertEqual(len(errors), 0, errors)
+
+    def test_empty_string_attribute(self):
+        """Verify that we can determine dtype for empty string attribute during validation"""
+        builder = GroupBuilder('my_bar',
+                               attributes={'data_type': 'Bar', 'attr1': []},  # <-- Empty string attribute
+                               datasets=[DatasetBuilder(name='data', data=['text1', 'text2'],
+                                                        attributes={'attr2': [10, ]}),
+                                         DatasetBuilder(name='dataInt', data=[5, ])
+                                         ])
+        self.runBuilderRoundTrip(builder)
+
+    def test_empty_string_dataset(self):
+        """Verify that we can determine dtype for empty string dataset during validation"""
+        builder = GroupBuilder('my_bar',
+                               attributes={'data_type': 'Bar', 'attr1': ['text1', 'text2']},
+                               datasets=[DatasetBuilder(name='data',    # <-- Empty string dataset
+                                                        data=[],
+                                                        dtype='text',
+                                                        attributes={'attr2': [10, ]}),
+                                         DatasetBuilder(name='dataInt', data=[5, ])
+                                         ])
+        self.runBuilderRoundTrip(builder)
+
+    def test_empty_int_attribute(self):
+        """Verify that we can determine dtype for empty integer attribute  during validation"""
+        builder = GroupBuilder('my_bar',
+                               attributes={'data_type': 'Bar', 'attr1': ['text1', 'text2']},
+                               datasets=[DatasetBuilder(name='data', data=['text1', 'text2'],
+                                                        attributes={'attr2': []}  # <-- Empty integer attribute
+                                                        ),
+                                         DatasetBuilder(name='dataInt', data=[5, ])
+                                         ])
+        self.runBuilderRoundTrip(builder)
+
+    def test_empty_int_dataset(self):
+        """Verify that a dataset builder containing an array of references passes
+        validation after a round trip"""
+        builder = GroupBuilder('my_bar',
+                               attributes={'data_type': 'Bar', 'attr1': ['text1', 'text2']},
+                               datasets=[DatasetBuilder(name='data', data=['text1', 'text2'],
+                                                        attributes={'attr2': [10, ]}),
+                                         DatasetBuilder(name='dataInt', data=[], dtype='int')  # <-- Empty int dataset
+                                         ])
+        self.runBuilderRoundTrip(builder)